requirements.txt
flask
flask-cors
requests
pandas
openpyxl
python-dotenv

boto3
qrcode[pil]
